[FEAT]: #09 - 클리어 유무에 따른 스테이지 grayScale 적용
- StageView.swift에 정의되어 있던 Item 모델을 따로 분리했습니다.
- ACarousel.swift, ACarouselViewModel.swift에 grayScale 적용하기 위한 코드 추가
- StageView.swift에 있던 불필요한 요소 제거
